+++ src/ILP/ilp-client.ts
@@ -1,8 +1,8 @@
-import { BigInteger } from 'big-integer'
 import { IlpClientDecorator } from './ilp-client-decorator'
 import DefaultIlpClient from './default-ilp-client'
-import { GetBalanceResponse } from '../generated/web/ilp/get_balance_response_pb'
-import { SendPaymentResponse } from '../generated/web/ilp/send_payment_response_pb'
+import { AccountBalance } from './model/account-balance'
+import { PaymentResult } from './model/payment-result'
+import { PaymentRequest } from './model/payment-request'
 
 class IlpClient {
   private readonly decoratedClient: IlpClientDecorator
@@ -25,34 +25,25 @@ class IlpClient {
   public async getBalance(
     accountId: string,
     bearerToken?: string,
-  ): Promise<GetBalanceResponse> {
+  ): Promise<AccountBalance> {
     return this.decoratedClient.getBalance(accountId, bearerToken)
   }
 
   /**
    * Send the given amount of XRP from the source wallet to the destination address.
    *
-   * @param amount A `BigInteger`, number or numeric string representing the number of drops to send.
-   * @param destinationPaymentPointer the payment pointer to receive funds
-   * @param senderAccountId the ILP account sending the funds
+   * @param paymentRequest A PaymentRequest with options for sending a payment
    * @param bearerToken Optional auth token. If using node network client, bearerToken must be supplied, otherwise
    *        it will be picked up from a cookie.
-   * @returns A promise which resolves to a `SendPaymentResponse` of the original amount, the amount sent
+   * @returns A promise which resolves to a `PaymentResult` of the original amount, the amount sent
    *        in the senders denomination, and the amount that was delivered to the recipient in their denomination, as
    *        well as if the payment was successful
    */
   public async sendPayment(
-    amount: BigInteger | number | string,
-    destinationPaymentPointer: string,
-    senderAccountId: string,
+    paymentRequest: PaymentRequest,
     bearerToken?: string,
-  ): Promise<SendPaymentResponse> {
-    return this.decoratedClient.sendPayment(
-      amount,
-      destinationPaymentPointer,
-      senderAccountId,
-      bearerToken,
-    )
+  ): Promise<PaymentResult> {
+    return this.decoratedClient.sendPayment(paymentRequest, bearerToken)
   }
 }